Don't update query log filters if user is selecting text
This likely isn't the proper way to do this, but a rough outline to smooth the user experience by not requiring holding alt or ctrl to select text. On desktop this appears to work, I'm not sure how mobile behaves, this may break things.

Edit as needed, this currently checks window.getSelection() on mousedown and again on click.

Related to: #2662

Signed-off-by: Alex Anderson <anderson.john.alexander@gmail.com>